+++ src/main/java/org/opensearch/knn/index/codec/params/KNNVectorsFormatParams.java
@@ -26,6 +26,7 @@ public class KNNVectorsFormatParams {
     private boolean hierarchyEnabled;
     private Function<Integer, Integer> numberOfSubspacesPerVectorSupplier;
     private final SpaceType spaceType;
+    private boolean leadingSegmentMergeDisabled;
 
     public KNNVectorsFormatParams(final Map<String, Object> params, int defaultMaxConnections, int defaultBeamWidth) {
         this(
@@ -58,6 +59,7 @@ public class KNNVectorsFormatParams {
         initHierarchyEnabled(params, defaultHierarchyEnabled);
         initNumberOfSubspacesPerVectorSupplier(params);
         this.spaceType = spaceType;
+        initLeadingSegmentMergeDisabled(params, KNNConstants.DEFAULT_LEADING_SEGMENT_MERGE_DISABLED);
     }
 
     public boolean validate(final Map<String, Object> params) {
@@ -119,5 +121,13 @@ public class KNNVectorsFormatParams {
             return;
         }
         this.numberOfSubspacesPerVectorSupplier = JVectorFormat::getDefaultNumberOfSubspacesPerVector;
+    }
+
+    private void initLeadingSegmentMergeDisabled(final Map<String, Object> params, boolean defaultLsmDisabled) {
+        if (params != null && params.containsKey(KNNConstants.METHOD_PARAMETER_LEADING_SEGMENT_MERGE_DISABLED)) {
+            this.hierarchyEnabled = (boolean) params.get(KNNConstants.METHOD_PARAMETER_LEADING_SEGMENT_MERGE_DISABLED);
+            return;
+        }
+        this.leadingSegmentMergeDisabled = defaultLsmDisabled;
     }
 }